Feature/discover screen (#9)
* fix navigation

* create new grid container for sharing between screens

* fix versions, and mappers.

* fix versions, and mappers.

* fix navigation routes.

* fix navigation routes.

* create genre module

* adding genre logic, new icons etc, make new browse logic

* adding genres to filters with logic and update ui

* using mvi instead of mvvm

* using mvi instead of mvvm

* remove not needed function again

+++ discover/src/commonMain/kotlin/com/kmp/movieapp/discover/data/model/response/DiscoverMoviesDto.kt
@@ -15,7 +15,7 @@ internal data class DiscoverMoviesDto(
     @SerialName("backdrop_path")
     val backdropPath: String? = null,
     @SerialName("genre_ids")
-    val genreIds: List<Int>? = null,
+    val genreIds: List<Long>? = null,
     @SerialName("original_language")
     val originalLanguage: String? = null,
     @SerialName("original_title")
